convert homepage to ts
